requirements.txt
pydantic>=2.9.0
pydantic-settings>=2.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.39
alembic>=1.14.0
python-dotenv>=1.0.1
bcrypt==4.1.2
passlib==1.7.4
PyJWT>=2.10.1
passlib>=1.7.4
python-multipart>=0.0.19
python-dateutil>=2.9.0
email-validator>=2.3.0
python-jose[cryptography]>=3.4.0
PyMySQL>=1.1.0